Add Bigint

+++ codebook/6_Math/Combination.cpp
@@ -1,8 +1,11 @@
 struct Comb {
-    ll n; vector<Z> _fac, _invfac, _inv;
+    int n;
+    vector<Z> _fac, _invfac, _inv;
     Comb() : n{0}, _fac{1}, _invfac{1}, _inv{0} {}
-    Comb(ll n) : Comb() { init(n); }
-    void init(ll m) {
+    Comb(int n) : Comb() {
+        init(n);
+    }
+    void init(int m) {
         m = min(m, Z::getMod() - 1);
         if (m <= n) return;
         _fac.resize(m + 1);
@@ -18,24 +21,25 @@ struct Comb {
         }
         n = m;
     }
-    Z fac(ll m) {
+    Z fac(int m) {
         if (m > n) init(2 * m);
         return _fac[m];
     }
-    Z invfac(ll m) {
+    Z invfac(int m) {
         if (m > n) init(2 * m);
         return _invfac[m];
     }
-    Z inv(ll m) {
+    Z inv(int m) {
         if (m > n) init(2 * m);
         return _inv[m];
     }
-    Z binom(ll n, ll m) {
+    Z binom(int n, int m) {
         if (n < m || m < 0) return 0;
         return fac(n) * invfac(m) * invfac(n - m);
     }
-    Z lucas(ll n, ll m) { // Mod 要在 1E5 左右, 動態 Mint
+    Z lucas(int n, int m) { // Mod 要在 1E5 左右
         if (m == 0) return 1;
-        return binom(n % Z::getMod(), m % Z::getMod()) * lucas(n / Z::getMod(), m / Z::getMod());
+        return binom(n % Z::getMod(), m % Z::getMod()) *
+            lucas(n / Z::getMod(), m / Z::getMod());
     }
-} comb; // 注意宣告, 若要換模數需重新宣告
+} comb; // 若要換模數需重新宣告
